src/keyboard.cpp: same layout as squeekboard (MR 86)
Layers 2 and 3 had several duplicate keys, and the '}' key was missing.
Let's just use the same layout as squeekboard. Notably, this adds a
period key (".") next to the space key.

This layout has all the keys from both previous layouts, except for:

	∑, ∞, ≈

I think it's fine to make this change, since postmarketOS is still in
beta. If somebody happens to use these in their passwords, they could
install the telnet initramfs hook and change their password, if
necessary.