keyboard: Add 4th row on first two layers for numbers / symbols (MR 111)
This adds a fourth row at the top on the letter / capital letter
layers. On the letter layer, the row contains the numbers 0 through
9. On the capital letter layer, the row contains the symbols associated
with the number keys.

This mimicks the behavior on password entries on Android and allows for
easy access to common numbers / symbols without going deeper into the
layers.

The 4th row is displayed conditionally if and only if the vertical
space is sufficient to allow for keys with a height / width ratio of
at least one. Otherwise, the previously existing  three-row layout
is used.

Closes: #116