Use OpenGL ES driver by default and add CLI switch to opt out (MR 113)
This change makes osk-sdl use the OpenGL ES driver whenever it is
available. If it is not available, the app falls back to the default
driver.

Additionally, a new command line option -G / --no-gles was added that
allows to opt out of the behavior and fall back to the default driver
from the start.

Closes: #102