Enable support for key press haptic feedback (MR 115)
This enables haptic feedback when typing on the keyboard. Vibrations are
triggered on touch/mouse down events, which seemed more natural than
triggering it on the up events.

If it's not working, then the module for the haptic/vibration device may
not be enabled or available (e.g. when in initfs)

Closes #95