util: add isDirectFB helper function (MR 122)
This function uses several methods to try and determine if the video
backend is directfb or not. This allows working around directfb quirks
in the app by conditionally disabling features that don't work on
directfb.

In cases where SDL_Init hasn't been called yet, the SDL_VIDEODRIVER env
var is used to determine if the app was run with DirectFB. In most
(all?) cases, some other more-preferred backend will be attempted if
this var is unset (X11, Wayland, kmsdrm), so it seems like the only time
osk-sdl/sdl would try to use directfb is if it's explicitly set in this
env. That's the assumption made here, at least.