main: show tooltip with instructions when no pass has been entered (MR 132)
This adds a new tooltip that shows when the app starts, and when no text
is in the input box (including when all text is deleted), informing the
user to input a disk decryption passphrase.

The tooltip disappears as soon as the user starts entering characters,
and the normal password dots replace it in the input box.

I decided to use a tooltip in this way since there was not enough space
on the screen to fit it elsewhere when the screen height is small (e.g.
on the N900), and I think this looks rather nice.

fixes #60