luksdevice: fix pushing event to main thread after unlock attempt (MR 134)
This mechanism was broken, or not really working as intended. The empty
event was ignored by the main loop, and I think what was intended was
for this to trigger a render event (e.g. for displaying an error). This
is necessary when osk-sdl is running without animations, since the
animation loop pushes render events continuously into the queue and this
is will *not* happen when animations are intended to be disabled.