main: set wallpaper as draw color, not texture (MR 137)
This reduces the render time for a 2 render pass (using sw rendering) by
something close to 50%, as measured on a Librem 5 with SDL2 sw
rendering. RenderClear clears the screen to the wallpaper color, thereby
saving quite a lot of time.

The render while loop was measured as a whole with gettimeofday
(converted to msec):

Before:
        INFO: render loop: 360 ms
        INFO: render loop: 367 ms
        INFO: render loop: 367 ms
        INFO: render loop: 362 ms

After:
        INFO: render loop: 175 ms
        INFO: render loop: 165 ms
        INFO: render loop: 154 ms
        INFO: render loop: 167 ms