Reduce height of input box when displaying without keyboard (MR 140)
When no keyboard is used (-x), the input box is comically large on some
systems (PineBook Pro). This wraps the height of the input box to be
slightly more than the font size. It makes some gross assumptions about
font size and pixels, but seems to work on the two devices I tested on
(librem 5 and PBP)