Handle cases where text input events are still fired for hotkeys (MR 144)
The newly added C-u hotkey seems to be generating a text input event in
SDL on some systems (Pinebook Pro), while it did not on other systems
(thinkpad running Arch) in testing the earlier patch.

This returns from the text input handling when the event was fired
because C-u was held down.