Don't trigger render event when repeat delay is not met (MR 144)
This is a slight optimization that moves scheduling of a render event so
that it is only scheduled when a repeat character satisfies the repeat
delay, which is the only time a dot will be drawn on the screen for the
character. There's no reason to render things if the key repeats before
the delay since the character is discarded and nothing new is added to
the screen.