util: render font glyphs to represent passphrase on the screen (MR 151)
This simplifies the way passphrases are represented on the screen by
replacing the drawn circle/dot with any specified character from the
font in osk.conf.

As an added bonus, users can set the option in the config to "" to have
osk-sdl hide any passphrase dots on the screen, which is the same
behavior cryptsetup has. Onlookers can't easily see how long a
passphrase is.