test/test_functional: use -t / window mode when running osk-sdl in tests (MR 153)
I'm not sure why this is suddenly needed, but my best theory is that a
change in SDL2 2.0.16 results in osk-sdl not displaying correctly when
running on X in fullscreen mode. You can see this by running an X server
at a specific size, then running osk-sdl inside of it.

$ Xephyr -screen 480x800 :11 &
$ SDL_VIDEODRIVER=x11 DISPLAY=:11 osk-sdl -d a -n a -c ./osk.conf

Previously it would fit in the X screen size specified, but now it seems
to resize the X screen and fill more space. This throws off all of the
tests, which all depend on using xdotool to move a cursor and click at
various coordinates in the X screen.